* added some support for 64bit feature: extended struct ext2_block_group, block number types changed from uint32 to off_t * added error traces, asserts * BitmapBlock::CheckUnmarked() and CheckMarked() computed a wrong remainingBits and mask
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@39203 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/ext2/BlockAllocator.h
@@ -27,11 +27,11 @@ public:
 
 			status_t	AllocateBlocks(Transaction& transaction,
 							uint32 minimum, uint32 maximum, uint32& blockGroup,
-							uint32& start, uint32& length);
+							off_t& start, uint32& length);
 			status_t	Allocate(Transaction& transaction, Inode* inode,
-							off_t numBlocks, uint32 minimum, uint32& start,
+							off_t numBlocks, uint32 minimum, off_t& start,
 							uint32& length);
-			status_t	Free(Transaction& transaction, uint32 start,
+			status_t	Free(Transaction& transaction, off_t start,
 							uint32 length);
 
 			uint32		FreeBlocks();
